New folder `db`. Separated DB class from update_db.py
