+++ settings.py
@@ -1,0 +1,1 @@
+DB_PATH = "db/db.json"
